refactor: make sigplot plugins available under sigplot.plugins
Historically, plugins were distributed as a second independent file.
This was done to minimize the size of SigPlot.  As SigPlot transitions
to an ES6 module syntax, and tools like WebPack and Browserify allow
developers to import individual SigPlot modules, this separation of
files complicates deployment and testing.  This refactoring includes all
basic plugins under `sigplot.plugins`.

The legacy plugin file is still produced and, while redundant, should
not break anything if included.