fix: incorrect rendering of draw_image
The draw_image takes a source buffer (bounded by xmin/xmax/ymin/ymax)
and selects some or all of the elements out of the source buffer based
upon the current plot view box.  In other words, if the view box bounds
exceed the source buffer bounds then the entire image is drawn; but if
the view box bounds are smaller than the source buffer bounds than a
subset of the source buffer is drawn.

When a subset of the source buffer is drawn, the selected source buffer
elements need to exceed the view box because there may be a fractional
part of the source buffer 'pixel' to be rendered.  Because the
underlying renderImage does not directly support sub-pixel rendering
we expand the source buffer to exceed the bounds of the view box and
then use clipping to avoid rendering the overdraw.